Extract common AsyncSave from adapters
